Update YAAAAT_apk_ripper.py
Those savvy enough to notice the version regressed deserve a cookie. The REGEX is broken and the programmatic stepping of the extraction of flagged strings is horribly inefficient. I did finalize some checks to at least let you run the code in either Python 3 and Python 2 but most of the functionality with REGEX need to be ported from 2 -> 3 so yeah :(.